# ARROW-12330: [Developer] Restore values at counters column of Archery benchmark