--- a/dev/archery/archery/benchmark/codec.py
+++ b/dev/archery/archery/benchmark/codec.py
@@ -50,6 +50,7 @@ class BenchmarkCodec:
             "values": b.values,
             "time_unit": b.time_unit,
             "times": b.times,
+            "counters": b.counters,
         }
 
     @staticmethod